Update checked and fixed all found offences
