+++ musicAlbum/genre.rb
@@ -1,3 +1,5 @@
+require_relative './musicalbum'
+
 class Genre
   attr_accessor :items
